feat(nns-recovery): download pool from another node using backup access (#6647)
Currently, we assume during an NNS recovery that the node owned by
DFINITY will have all required consensus artifacts until the highest
replay height, but this might not be the case.

This PR adds a new step to `ic-recovery` to allow the coordinator to
download the consensus pool from a different node than the one they're
downloading the state from (which is the DFINITY-owned node, because it
requires admin access). This is possible because the consensus pool is
readable with the backup access. In a real recovery, we would arguably
select this node as the one with highest finalization height, [similarly
to current application subnet
recoveries](https://github.com/dfinity/ic/blob/590c15bc2f45003f51bec9deb49ae69135333c93/rs/recovery/src/app_subnet_recovery.rs#L40-L43).

In the NNS recovery system tests, this implies that we do not need to
rely on the DFINITY-owned node to be the one with highest certification
share height anymore. These changes also allow to add the
`nr_broken_dfinity_node` system test that explicitly chooses the
DFINTY-owned node as faulty.