feat(crypto): CRP-2912 in-memory canister secret key store (#6496)
Adapts the Crypto Vault so that the canister secret key store is purely
memory-based and the data in that store is no longer persisted to disk
upon changes. This is to improve the performance of threshold
signatures.

Because the data is no longer persisted to disk, the data loaded in the
canister secret key store by Consensus does not survive node restarts.
However, this is still safe because Consensus currently always loads
transcripts "on demand" in the following two places:
1. when a signature share is created (then the key transcript and all
pre-signature transcripts are loaded)
2. when a dealing for a transcript is created (then all dependencies of
that transcript are loaded, i.e. kappa/lambda or lambda/key for the
multiplication in ECDSA pre-signatures, or the reshared transcript
during resharing)

Also, the IDkg transcript corresponding to the master keys is part of
every summary block/catch-up package, which means it will also exist on
our backup pods.

The PR also ensures that the existing canister secret key store file is
cleaned up, i.e., zeroized and deleted, upon startup of the Crypto Vault
process.

The change is also downgrade compatible: upon a downgrade, the canister
secret key store file won't be present any more (because the changes in
this PR deleted it), but the Crypto vault will create a new, empty key
store file, and this will be populated when the Consensus layer loads
the transcript on demand (see above). Further, the
`upgrade_app_subnet_test` and the `upgrade_downgrade_app_subnet_test`
system tests also ensure that after upgrade/downgrade, chainkey
signatures are still working.

The tests for the in-memory SKS were copied from the protobuf-based SKS
and adapted.

Note that this change only affects the canister secret key store, but
not the node secret key store.

---------

Co-authored-by: Leo Eichhorn <99166915+eichhorl@users.noreply.github.com>