chore(crypto): CRP-2928 Fix some formatting errors introduced in Rust 2024 Edition conversion (#6798)
Previously the Rust formatting rules incorrectly assumed that any
comments on a line following an item with a trailing comment should
always be indented to match the trailing comment.

This is fixed in the Rust 2024 edition; now the formatter incorrectly
assumes that any comments on a line following an item with a trailing
comment should never be indented to match the trailing comment.

Progress.