fix(btc-adapter): improve btc-adapter responsiveness (#6760)
Fix XC-476

Improve the runtime behavior of btc-adapter by making blocking
operations async, including:
- `to_socket_addrs` may involve DNS lookup
- `add_header` may involve disk I/O, and also header validation could be
time consuming too.

As a result, the gRPC server run by the adapter is now more responsive
and less prone to timeouts.

`BlockchainState` has to be made interior mutable in order to support
async `add_header`.
The use of `Mutex` over `BlockchainState` is no longer needed.

This PR also cleaned up the use of tokio handlers, and fixed a problem
that caused the integration
test to be flaky.