refactor(governance): Use slightly more authoritative source of Subnet Rental canister API types. (#6778)
# Related Work

In [PR 6722], I tried sourcing the definition of these types from the
the true authoritative source, the [subnet-rental-canister repo][src].
Unfortunately, that created various headaches. The headache that
convinced me to abandon the aforementioned PR was how it increased the
size of the NNS Governance WASM. The increase was not super big, but
since WASM size is a more or less scarce resource, I didn't want to eat
into our remaining headroom. Therefore, I went for this less ambitious
PR instead.

[PR 6722]: https://github.com/dfinity/ic/pull/6722

[src]: https://github.com/dfinity/subnet-rental-canister

# References

Closes [NNS1-3965].

[NNS1-3965]: https://dfinity.atlassian.net/browse/NNS1-3965