perf(crypto): CRP-2918 Add variable-time batch inversion to tECDSA
This takes advantage of "Montgomery's trick" to perform many inversions using
just a single inversion calculation, plus some extra multiplications.
Benchmarks show that using a batch inversion even for 32 elements is faster than
just 2 variable-time inversions.

Modify the calculation of the Lagrange coefficients to use the batch inversion.
This reduces the cost of computing the Lagrange coefficients as follows on a
test machine, for n = 34 (matching the production signing subnet configuration).

* secp256k1 from 380 μs to 91 μs
* secp256r1 from 617 μs to 109 μs
* ed25519 from 455 μs to 168 μs