chore: use a function instead of a for-loop in bazel-test-all (#6792)
We would like to start doing stuff between the `bazel build` and `bazel
test` like uploading images to our bazel cache and uploading artifacts
to S3 (in parallel with `bazel test`). To make this easier to implement
we replace the for loop over `build` and `test` with a `bazel_do`
function which is applied to `build` and then to `test`.

---------

Co-authored-by: IDX GitHub Automation <>