test(networking): State sync performance test (#6657)
This PR adds a new state sync performance test to the system tests.

The test:
- Is colocated and uses limited bandwith and latency
- Creates a single node subnet with NNS and multiple state sync
canisters
- Increases the state of the canisters
- Then proposes to let 12 nodes join the subnet
- Measures how long the 12 nodes need to state sync

Currently, some test parameters are hardcoded, but might be made
variable in future tests:
- 300 Mb links with 150ms latency
- 8 GiB state

In this scenario, the initial state sync usually takes around 120
seconds

The test reuses a lot of code from the `rejoin_test_lib` library. Some
of the code in this library was adapted by:
- making some function public
- Returning some values that prior where internal
- Adapting some timeout values
- Reducing output verbosity

---------

Co-authored-by: kpop-dfinity <125868903+kpop-dfinity@users.noreply.github.com>
Co-authored-by: Bas van Dijk <bas@dfinity.org>