feat(btc-adapter): implement a hybrid header cache (#6797)
Implement a hybrid header cache that would persist anchor header and its
ancestors to disk.
This leads to a simpler on-disk cache implementation, since it only has
to persist a single chain of headers.
The downside is that headers since the anchor point would be lost if the
adapter is restarted, in which case it would sync more headers in order
to catch up.

---------

Co-authored-by: mducroux <mathieu.ducroux@dfinity.org>