chore: do not check memory allocation bounds when loading checkpoint (#6840)
This PR introduces a new (temporary) function
`MemoryAllocation::new_unchecked` skipping the check of an upper bound
(506 GiB) on canister memory allocation when loading a checkpoint.

This change is the first stage of an effort to drop the upper bound
completely: first (this PR) we ensure that checkpoints with arbitrary
memory allocation can be loaded while still validating user requests to
not set arbitrary memory allocation. After this PR is rolled out to all
subnets, we disable the check when validating user requests, too. This
way, it is guaranteed that downgrades of a replica version containing
the follow-up [PR](https://github.com/dfinity/ic/pull/6821) are possible
(due to this PR).