test(nns-recovery): add local NNS recovery test (#6515)
This PR adds a variant to the NNS recovery test to execute the recovery
"locally", i.e. it runs the recovery tool directly on the DFINITY-owned
node.

Some functionality has been improved:
- [CON-1584] Allow the recovery coordinator to set `upload_method`
during the `UploadState` step to leave blank during
`UploadCUPAndRegistry` and fill it in `UploadState`. This allows to
upload the state to the node but not the recovery artifacts, in case we
want to fix this node like the other NPs through the recovery
upgrader/engine.
- Add a `wait_for_cup_node` parameter to wait for the CUP on the local
node, similarly to app subnet recoveries
([source](https://github.com/dfinity/ic/blob/24b0dc2cacc5841866c4cb26cad3dd0e84e40a94/rs/recovery/src/app_subnet_recovery.rs#L145-L147)).
This is because the CUP is not served on `::1` (localhost), so this step
needs the public GuestOS IP even for a local recovery (we cannot just
rely on `upload_state_method` because it could be `local`).
- [CON-1572] The `DownloadState` step also follows what is done in app
subnet recoveries
([source](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/recovery/src/app_subnet_recovery.rs?L359-374))
and uses `cp` instead of using `rsync` on `::1`, which is significantly
faster for that amount of data.
- The output directory for recovery artifacts can be set at runtime (as
a prompt to the user).
- As the `WaitForCUP` step upgrades the node, we need the recovery
artifacts to be stored in a shared mount of the node, otherwise the
recovery artifacts would be lost because of the partition swap. The
default location is thus now `/var/lib/ic/data/recovery/output` and some
docs/warnings are printed. Note that this means that they will now get
removed by the `Cleanup` step and the coordinator should thus export
them prior to that.
- As a consequence, we also add a `--skip` parameter to skip the
`Cleanup` step in the local recovery system test.

In the system test, we send the SSH keys to the remote node (via
`scp_send_to`) and run the local recovery. We then fetch the recovery
artifacts (via `scp_recv_from`).

[CON-1572]:
https://dfinity.atlassian.net/browse/CON-1572?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[CON-1584]:
https://dfinity.atlassian.net/browse/CON-1584?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ