fix(nns-recovery): break DFINITY-owned node last (#6839)
This PR reduces the flakiness of
`//rs/tests/nested/nns_recovery:nr_broken_dfinity_node`.

When we break the DFINITY-owned node, there is a small chance that,
during the time where we break nodes one by one, the subnet still makes
progress and reaches the next DKG interval. If that is the case, then
the DFINITY-owned node would have a state from the previous interval and
we would not have the required consensus artifacts to replay until the
latest certified height.

Example of a failed run: With a DKG interval size of 49, 1400 is a CUP
height. We see [at this
line](https://dash.zh1-idx1.dfinity.network/invocation/a6f2f5a1-5afb-4c28-87c1-186754d8ec99?target=%2F%2Frs%2Ftests%2Fnested%2Fnns_recovery%3Anr_broken_dfinity_node_head_nns&targetStatus=11#@1732)
that the highest certification share is 1403, just above the CUP height.
During the `ICReplay` step
[here](https://dash.zh1-idx1.dfinity.network/invocation/a6f2f5a1-5afb-4c28-87c1-186754d8ec99?target=%2F%2Frs%2Ftests%2Fnested%2Fnns_recovery%3Anr_broken_dfinity_node_head_nns&targetStatus=11#@1960),
we do not have the state at height 1400 and are stuck here until the
test times out. Probably the DFINTY-owned node was broken just before
height 1400 and the subnet continued for a few heights.

To fix this, this PR breaks the DFINITY-owned node last instead of
first. With this setup, there is no way the DFINITY-owned node can be
one DKG interval behind the others.

Note that this edge-case was actually the reason why we went for
`LARGE_DKG_INTERVAL` (=49) for this step, to reduce the chances of that
scenario happening. We can thus revert to `DKG_INTERVAL` (=9) to run the
test slightly faster.

P.S.: If this were to happen on mainnet on all three of the
DFINITY-owned nodes, that's where we would use the state present on
backup pods to perform the recovery.