chore: colocate //rs/tests/dre:guest_os_qualification (#6883)
The `//rs/tests/dre:guest_os_qualification` test has a flakiness rate of
over 10% from last week onwards (2025-09-16 - 2025-09-23). Some of the
flakiness comes from `No route to host` errors when trying to fetch
metrics from IC nodes. To reduce the chance of networking failures
interfering with the test this colocates the driver with the testnet it
will deploy.

Future work is to add some retry logic around the fetching of metrics.

---------

Co-authored-by: IDX GitHub Automation <>