fix(ICRC-Rosetta): FI-1868: fix exponential backoff in case of an error (#6734)
Currently, if any of the syncing operations succeeds, the backoff is
reset. I.e. if we fail to sync the new block but we compute the balances
correctly (since we did not add any new blocks), the backoff would not
be applied, even though rosetta is permanently stuck on an error.