test: use supermicro instead dell in tecdsa_performance_test system subnet (#6913)
Adapts the tecdsa_performance_test template so that the system subnet
uses the Supermicro host feature instead of Dell.

With this, we will be able to run the test for 34 nodes, given that the
Farm testing infrastructure has exactly [34 Dell performance
hosts](https://github.com/dfinity-lab/farm/blob/master/dcs.tf#L198-L231).

The Supermicro hosts are known to be slower than the Dell ones, but it
should not make a difference for the test if the system subnet uses
these slightly slower hosts because the expensive work is done by the
application subnet, which still has the Dell nodes.