fix(PocketIC): zombie process after panic (#6912)
This PR fixes the PocketIC server to not become a zombie process after
an operation panicked.

So far, the PocketIC server increments a counter before executing an
operation and decrements the counter after the operation executed.
However, if an operation crashes, then the counter is not decremented
and the server believes that the (failed) operation is still pending.
This PR fixes that issue by using `Drop` guards that are guaranteed to
run even after panics.

Since `Drop` guards are synchronous, this PR switches from
`tokio::sync::RwLock` to `std::sync::AtomicU64` for storing the server's
TTL.

The issue has been reproduced and the fix tested manually by introducing
an artificial panic in the server and running:
```
./pocket-ic-server --ttl 3 --port 8080
```
in one tab and
```
curl http://localhost:8080/api.json
```
in another tab. Before the fix, the server will not stop after its TTL
of 3s. After the fix, the server terminates after its TTL of 3s.