chore(ICRC_Rosetta): FI-1871: Export relevant block_height metrics (#6896)
Currently, ICRC Rosetta exports a `rosetta_verified_block_height`
metric, corresponding to the verified tip of the ledger.

This PR proposing the following changes:

- Change the value of the `rosetta_verified_block_height` metric to
correspond to the block height for which Rosetta has synced and verified
the blocks.
- Export a `rosetta_synched_block_height` metric corresponding to the
index of the most recent block that Rosetta has synced and stored.
- Export a `rosetta_target_block_height` metric corresponding to the tip
of the ledger (previously exported as `rosetta_verified_block_height`).

For synced Rosetta instances, the three metrics above will mostly have
the same value. When syncing from scratch, or after a longer downtime,
there may be larger differences. By e.g., comparing the
`rosetta_target_block_height` metric (the height of the ledger) and the
`rosetta_verified_block_height` metric, it is possible to determine if
Rosetta is making progress.