test: Add migration canister to system test framework (#6871)
This PR adds the migration canister to the system test framework, so
that when a system test installs all NNS canisters, the migration
canister is installed along. At the same time, it adds it to the
expected list of canisters for tests that start with a golden NNS state
(i.e. the mainnet state), and introduces a basic canister migration
system test that does interacts with the migration canister. The new
test is still very bare-bone, but will be expanded on later.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>