feat: Implement SEV disk encryption key exchange (#6576)
The PR implements the following 4 components which are each implemented
in a separate crate (ordered by the recommended review order):
- shared
Contains shared utilities, the protocol definition and the data
structure for calculating the custom data in the SEV attestation report.
- server
The server implementation that will run in the VM that we're upgrading
from (aka old VM). The implementation is fairly convoluted because
extracting TLS keys from the connection had to be implemented separately
(not supported by the simple out-of-the-box server). The implementation
has the following components:
- Agent: high-level API that will be called from the orchestrator (not
in this PR yet), sets up the service, the server and launches the
Upgrade VM using the VSOCK interface
  - Service: implementation of the key exchange protocol
- Server: setting up a listener, accepting connections and extracting
TLS public keys
- client
The client implementation (planned to be run as a separate binary). The
client is also somewhat complicated because it also has to extract the
server's TLS public key.
- tests
The tests run the protocol E2E by bringing up a server, mocking the
VSOCK interface and launching a client.

Some things worth noting:

- We disable certificate checks during the TLS handshake since both the
client and the server use self-signed certificates.
- I had to change the way we disable SEV root certificate checks in
tests. Previously, this way enabled under `#[cfg(test)]` but this config
is only active in a crate's own tests so it couldn't be used from the
newly added crates. The new approach uses a newly added
`SevRootCertificateVerification` enum.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>