feat(sns): add network to RegisterExtensionArgs (#6938)
Add the field `network` as dfx doesn't handle extension arguments
correctly.

Currently, the following command fails:

```bash
dfx sns propose-to-register-extension \                                         
  --sns-root-canister-id lacdn-3iaaa-aaaaq-aae3a-cai \
  --wasm-path ~/javad.wasm.gz \
  --proposal-url https://example.com \
  --summary test \
  --network ic
```

as dfx doesn't pass the argument `--network` to the extension. As a
workaround, we provisionally added a field called `network` to
`RegisterExtensionArgs`, so that `--network ic` is no longer treated as
input to the dfx but to the extension.