feat: Allow loading a snapshot onto another canister (#6592)
This PR lifts the restriction that a canister snapshot can only be
loaded on the canister that took it. A snapshot can now be loaded onto
another canister, of course only in the case that the two canisters are
on the same subnet and the sender is the controller of both canisters
(the one owning the snapshot as well as the target).

---------

Co-authored-by: mraszyk <31483726+mraszyk@users.noreply.github.com>