fix(PocketIC): new endpoint to prune result of long-running computation (#6988)
This PR adds a new DELETE endpoint `/prune_graph/<state_label>/<op_id>`
to the PocketIC server for pruning the result of a long-running
operation.

This endpoint should be called after successfully reading the result
using the GET endpoint `/read_graph/<state_label>/<op_id>`. The
`state_label` and `op_id` are returned by `ApiResponse::Started
{state_label, op_id}`.

If the new endpoint is not called, then the result of a long-running
computation is stored in the PocketIC server forever, resulting in a
memory leak.