test(node): remove recovery_upgrader_test (#7005)
NODE-1707

The guestos-recovery-upgrader logic is already covered by the more e2e
`nns_recovery` tests. The recovery upgrader test was created as a
building block to get to these end to end tests.

When the test was originally created, it used the mainnet guestos update
image releases as the image upgraded to, but after we started publishing
recovery-GuestOS images and changed the recovery-GuestOS download file
path to the recovery-GuestOS variant, we no longer could use mainnet
guestos update image releases, and we disabled this test.

We could update this test to use recovery-GuestOS update images, which
would mean the test would download from the public CDN, which we don’t
want. Or we could impersonate the upstreams and setup an upstream VM to
host the recovery-GuestOS image (which is what we do in the e2e recovery
tests). Or, we could just remove this test. I’ve decided to just remove
the test, as I don’t think this coverage is worth the time to implement
or complexity it adds, as we already have this functionality tested
extensively by the `nns_recovery` tests, and it's not a particularly
complex piece of logic.