test(service-nervous-system): Avoid spurious swap upgrade test failures. (#6919)
There are two measures for 👆:

1. In test, add cycles to canisters before trying to upgrade them. This
is to avoid test failures due to lack of cycles, which has nothing to do
with what we are trying to test, i.e. that swap is upgradeable.

2. Remove a couple of SNSs from the list of swaps that need to be
tested. This is fine, because their swap failed, and so we do not care
if it can be upgraded.
    1. Juno Build
    2. MORA DOA