fix(http-outcalls): Make invalid_ip correctness test less flaky. (#6989)
**Before**:
The test was asserting a response that it got from calling out to
`240.0.0.0`.
This had several issues:
- when API Boundary Nodes will be used as socks proxies for the app
subnet, they won't have a static IPv4 address anymore
- the test was asserting for a timeout (which does happen in 97% of the
cases), however in the remaining 3%, the request fails mot because of a
timeout, but because the "[Time to
live](https://en.wikipedia.org/wiki/Time_to_live)" expires before the 30
seconds deadline (which is non deterministic). This made the error type
flaky (either `SysFatal` or `SysTransient`).

**After**:
The test is calling out to `2001:db8::1`, which is a [reserved IPv6
address](https://en.wikipedia.org/wiki/Reserved_IP_addresses#IPv6). It
now consistently fails with `SysTransient`.

---------

Co-authored-by: kpop-dfinity <125868903+kpop-dfinity@users.noreply.github.com>