fix(ic-recovery): `ic-recovery` fixes and improvements (#6893)
This PR does quite a lot of changes to `ic-recovery`, but most of them
are quite small in size and I thought it would have been overkill to
split them all in separate PRs. Let me know if you think otherwise. 1
commit == 1 change, so I would recommend reviewing by going through each
commit instead of browsing the files changed:
-
https://github.com/dfinity/ic/pull/6893/commits/1c207957d66f80c89ec7e592cc281987cf7ef403
(bugfix): Because of `key_file: alt_key_file.or(self.key_file.clone())`
in `DownloadCertificationsStep`, if we had an admin SSH key file and not
a backup SSH key file, we would incorrectly use the admin SSH key
instead of using the agent for NNS recovery on same nodes.
-
https://github.com/dfinity/ic/pull/6893/commits/26e3204021e1843a059ab166f0efed34a262e9fc
(bugfix): `DownloadIcStateStep` was always using the admin SSH key file,
meaning we would incorrectly use it also when downloading the consensus
pool.
-
https://github.com/dfinity/ic/pull/6893/commits/d3736734977a533dbdbf4a8776143dede14a5cc6
(refactoring): In NNS recovery, instead of having a dedicated step to
back up the state (`CopyIcState`) we use the `keep_downloaded_state`
argument during `DownloadIcStateStep`, which was introduced for
application subnet recoveries but I guess was overlooked to be applied
also to NNS recoveries. In the new implementation, we won't ask the user
whether to keep a copy of the state if the recovery is done locally.
This is because a copy of the state is anyways at `/var/lib/ic/data`,
i.e. the "live" location. It can still be forced by explicitly setting
the CLI argument `--keep-downloaded-state` though. This follows the
exact same implementation as app subnet recoveries.
-
https://github.com/dfinity/ic/pull/6893/commits/d5502c27b2e4a91c842897fb349a0380e1929204
(feature): If `--keep-downloaded-state` is already explicitly set in the
CLI, do not ask the user again.
-
https://github.com/dfinity/ic/pull/6893/commits/99ff84ed794fbed7cedeee3b2eb07828175a3ddf
(feature): Similarly, if `--wait-for-cup` is already explicitly set in
the CLI, it will take precedence over `--upload-method` in a remote
recovery.
-
https://github.com/dfinity/ic/pull/6893/commits/08012e9a783cec2af4dca8bdab57a8fba48d233c
(feature): For safety, this change performs the state upload *before*
uploading the CUP & registry local store to make sure we do not lose
access to the node in case of a faulty reboot.
-
https://github.com/dfinity/ic/pull/6893/commits/082fc99b961cacbf15787250b1993b7b5e5298d6
(refactoring): In application subnet recoveries, the `download_method`
and `upload_method` are conceptually very different: the former is the
highest finalized node (with readonly access) while the second is a
DFINITY-owned node (with admin access). In NNS recovery, the
`download_state_method` and `upload_method` are effectively the same
DFINITY-owned node with admin access. This commit removes the
distinction between the two.
-
https://github.com/dfinity/ic/pull/6893/commits/2642c5344728619ada94b81d1711f7cc61f28b38
(feature): Display the default output directory during
`CreateArtifacts`.
-
https://github.com/dfinity/ic/pull/6893/commits/ea3743a6419a8d74b5f35883f87e34188b0e954c
(feature): Display progress of the download of certifications (through
`{i+1}/{n}`) and abort the program if less than `n - f` nodes are
reachable, as this would break the assumed hypothesis that at most `f`
nodes are corrupted/offline. Finally, it also tries to download
certifications from other nodes even if creating a directory fails
(`return success;` -> `continue;`) just in case it somehow works for
following nodes.
-
https://github.com/dfinity/ic/pull/6893/commits/fdb1bfe194d0afb4884b06a9273b626f616265d5
(nit fix): Since [this PR](https://github.com/dfinity/ic/pull/6818),
`--use-local-binaries` is ignored in NNS recoveries, so we can remove it
from the command in the system test.
-
https://github.com/dfinity/ic/pull/6893/commits/7117509c76246df6cb70d53e3a41b52ee6a87005
(refactor): Prefix `RecoveryArgs.key_file` with `admin_` and
`AppSubnetRecoveryArgs.pub_key` with `readonly_` to better indicate with
which SSH user they are intended to be used.
-
https://github.com/dfinity/ic/pull/6893/commits/21c7b3b8d4beb7e454c0365f5a7be65dcc92d829
(nit style): `DownloadRegistryStoreStep` reads a remote local store. It
does not necessarily need admin access so enlarge the API to accept more
SSH users/keys.
-
https://github.com/dfinity/ic/pull/6893/commits/9b4dca23f91e49acd05a2090c32c556f9ac7afc9
(chore): Increase NNS test total timeouts as the setup phase sometimes
take longer than 10 minutes
([example](https://dash.zh1-idx1.dfinity.network/invocation/6de11672-fa54-4406-b5f0-2332e925ed10?target=%2F%2Frs%2Ftests%2Fnested%2Fnns_recovery%3Anr_broken_dfinity_node_head_nns&targetStatus=11#@3084)).