fix(networking): Fix state sync performance test (#7007)
The state sync performance test has been problems getting stuck on
`await_status_is_healthy`.

It seemed to work occasionally, but often, it would get stuck, doing the
actual request to the `status` endpoint.
Looking at the logs, sometimes the request was actually made, while
other times it was not being made, and the request was not timing out
after 6 seconds.

I replaced the call with it's async version and now it seems to work. I
do not fully understand the previous behaviour, as I thought, calling
`block_on` while already being in an async block should be handled
gracefully.

This PR does not reactivate the test yet, as I would like to get a
better understanding first.