chore(governance): Delete the FulfillSubnetRentalRequest flag. (#6692)
When merge: after the Swiss subnet folks have their subnet. Nominally,
that is supposed to happen in mid Sep. However, there is a
non-negligible chance that things will be delayed. That event will
"probably" have taken place by October.

In any case, in the meantime, this is ready for review.

# References

Closes [NNS1-4115].

[NNS1-4115]: https://dfinity.atlassian.net/browse/NNS1-4115