feat(nns): Record known neuron votes before clearing ballots (#6848)
# Why

If a known neuron does not vote on a proposal, this fact should be
recorded (`vote = Vote::Unspecified`), and the only way to know that
they did not vote is to check the ballots after the voting window. One
way is to check the ballots right before they are cleared.

# What

- Record the votes by known neurons right before the ballots are
cleared.