feat: Add a split function to State Machine Tests that actually splits the Replicated State (#6897)
The previous way of simulating a subnet split was to just remove a
canister from the state and then import it into another `StateMachine`.
This doesn't use the actual subnet splitting code and leaves other parts
of the state entirely on the original subnet.

This PR adds a `split` function that essentially clones the
`StateMachine`, adapts the registry like a subnet split would and splits
the state on each side using the splitting logic directly. This should
more accurately represent the upcoming faster subnet splitting.

---------

Co-authored-by: Stefan Schneider <31004026+schneiderstefan@users.noreply.github.com>