perf(crypto): adapt vault message size histrogram buckets (#6916)
Adapts the histogram buckets for the crypto_vault_message_sizes from
* 1KB, 10KB, 100KB, 1MB, 2MB, 4MB, 8MB, 16MB, 20MB, 24MB, 28Mb, 30MB, to
* 0.5KB, 1KB, 5KB, 10KB, 50KB, 100KB, 500KB, 1MB, 4MB, 8MB, 16MB, 32MB

so as to get better resolution for message below 1MB, and less
resolution for large messages, which is not needed. We keep the amount
of buckets (12) the same, so as not to increase load on the monitoring
infrastructure.

For context, according to the [respective Grafana
dashboard](https://grafana.mainnet.dfinity.network/d/crypto-rpc-msg-sizes-and-ser-times/rpc-message-sizes-and-serialization-runtimes):
* Most requests are 0KB-1KB-10KB, except for
* `idkg_create_dealing` (0KB-1KB-10KB; 100KB-1MB), which will be adapted
by https://github.com/dfinity/ic/pull/6899 to also be 0KB-1KB-10KB
  * `idkg_load_transcript` (1KB-10KB-100KB-1MB)
  * `load_threshold_signing_key` (10KB-100KB-1MB)
  * `sign` (0-1KB-10KB-100KB) 
  * `threshold_sign` (0-1KB; 100KB-1MB-2MB-4MB-8MB-16MB)
* Most responses are 0KB-1KB, except for 
  * (NI-DKG) `create_dealing` (10KB-100KB) and
  * `idkg_create_dealing` (1KB-10KB)