feat: NNS Proposals to pause/unpause migrations
This commit introduces NNS proposals to call the `enable_api` and
`disable_api` endpoints on the canister migration canister. These
proposals will be called as part of subnet splitting to ensure that
subnet splitting and canister migrations do not overlap.

In addition, the commit tests the proposals by executing them as part of
the canister migration tests and modifies the migrations canister to
allow privileged calls from the governance canister.