perf(idkg): CON-1593 Use multi-threading in `validate_dealings` (#6962)
This PR makes use of the IDKG thread pool to validate IDKG dealings in
parallel. This required the following logic change:

Previously, we would iterate over all unvalidated dealings and validate
them one-by-one. While iterating, if we encountered a dealing by the
same dealer and transcript for which we already validated a dealing in
the current unvalidated batch, the second dealing by that dealer was
invalidated immediately.

With this PR we iterate over unvalidated dealings in parallel.
Therefore, filtering out duplicate (valid) dealings in the same batch is
impossible without some locking. For that reason we instead validate all
dealings of the current unvalidated batch optimistically, and then
filter out duplicate dealings after validation.

In addition, the following change was made for performance reasons:
Previously we would resolve the parameter reference first, and then
check if the dealing was created by a valid dealer and whether it is a
duplicate. However, this can also be determined based on the parameter
reference itself, without resolving it first, in which case the
validation fails sooner.

---------

Co-authored-by: Pierugo Pace <pierugo.pace@dfinity.org>