chore: Shorten NNS delegation refresh interval
This commit shortens the interval for subnets to refresh their
delegation from the NNS from 10 minutes to 5 minutes.

The upcoming canister migration feature is quite sensitive to outdated
NNS delegations, as ingress calls to a recently migrated canister are
only valid if the delegation reflects the updated routing table.

Canister migrations take at least 5 minutes anyway to protect from
replay attacks. It therefore makes sense to also set the delegation
update interval to 5 minutes. At the same time, updating delegations is
not so costly that we cannot do it every 5 minutes.