refactor(nns/sns): Use patching for test canister candid files (#6947)
# Why

Every time the candid file in NNS/SNS Governance needs to be updated,
the _test.did also needs to be updated, while the 2 files are mostly
identical.

# What

Use a `genrule` for the `_test.did` from `.did` and `_test.did.patch`,
while providing 2 helper targets when the updates are more complicated
(e.g. lines surrounding the patch are updated), in which case the
`_test.did` can be generated, manually updated, and the
`_test.did.patch` can be generated again.