fix: add empty governance_test.did files to make cargo clippy succeed (#7079)
https://github.com/dfinity/ic/pull/6947 removed `governance_test.did`
files and generated them with bazel instead. However these files are
included in rust code (`include_str!("governance_test.did")`) which
means that `cargo clippy` will
[fail](https://github.com/dfinity/ic/actions/runs/18238768918/job/51941823743#step:5:2903)
on it.

The reason the `cargo clippy` job
[succeeded](https://github.com/dfinity/ic/actions/runs/18230285300/job/51912091777)
on the PR was that the `cargo clippy` step only runs if `.rs` files have
been modified which the PR didn't touch.

To fix the `cargo clippy` job we commit empty `governance_test.did`
files.