test(recovery): use separate SSH keys in recovery system tests (#7009)
This PR enables the generation of new SSH keys for any user, not just
`admin`. It is used in recovery system tests as follows:
- NNS recoveries: create a separate SSH key for `backup` access.
Additionally, we remove `admin` access on all nodes except the
DFINITY-owned one to perfectly mirror the production setup.
- App subnet recoveries: create a separate SSH key for `readonly`
access. Unfortunately, we cannot blindly remove `admin` access like
above because the last check of the test (`assert_node_is_unassigned`)
uses admin SSH access. Thus, we still do so but we keep track of the SSH
connections that we opened to perform this operation and re-use those
for the last check. We still successfully prevent new `admin` SSH
connections from being established.