feat(nns): Allow updating known neuron through RegisterKnownNeuron (#7021)
# Why

Previously, a `RegisterKnownNeuron` proposal can still be made and
executed if the neuron is already a known neuron, as long as the "name"
is different. This is considered a hack because it technically allows
updating an existing a known neuron, while the name check was meant for
registering a new known neuron.

On the other hand, it's reasonable to allow "upsert" semantic for
RegisterKnownNeuron proposal type, and the recent addition of
`committed_topics` and `links` make this PR necessary - existing known
neurons would be able to update their information through proposals.

# What

* Relax the "known neuron name existence" check by not failing when both
the name and neuron id are the same.