feat: whitelisting mechanism for node swapping feature (#6710)
This PR adds a (temporary) whitelisting mechanism that will be used for
phased rollout of the node swapping feature. The whitelisting can be
enforced on two criteria:
1. Is the caller allowed to perform the node swapping? If no => stop,
2. If the caller is allowed, does he target an `old_node_id` that is
within a subnet which has this feature enabled? If no => stop.

Both of the flags have been added to the `RegistryCanisterInitPayload`
to ensure easy testing. The gradual rollout of this feature will be
performed by extending the **default values in
rs/registry/canister/src/flags.rs**. In this way the governance team
doesn't have to change their upgrade process.